feature(business-app): introduced graphql

+++ party/business/src/main/kotlin/cz/tyckouni/mvpgate/party/business/usecase/list/Page.kt
@@ -1,6 +1,9 @@
 package cz.tyckouni.mvpgate.party.business.usecase.list
 
+/**
+ * Page of entities
+ */
 class Page<T>(
     val elements: List<T>,
-    val totalSize: Int,
+    val totalSize: Long,
 )
